fix: 血量缩放链路防溢出与精度加固（生存模式词条预备）
- ApplyHealthMultiplier 倍率 float→double：消除血量 >2^24(≈1677万) 的 float 整数精度悬崖
- GetZombieHpMultiplier 同步 double，保持缩放链路无 float 截断
- 本体/头盔/护盾 6 处 *2/3 破碎阈值改 64 位中间量，避免 MaxHealth 极大时 *2 在 int 内溢出(约 >10.7 亿即翻负)
- 血量字段保持 int（21.4 亿空间足够，当前 5%/轮线性倍率下数值逐位不变，仅扩大极端值安全边界）

+++ PlantVsZombies/Game/Board.h
@@ -227,10 +227,10 @@ public:
 	// 目前唯一来源是生存模式（线性 1 + SURVIVAL_HP_GROWTH*(轮次-1)）；
 	// 未来其他模式（困难冒险、悬赏关等）若需血量倍率，在此处继续叠乘即可，调用方无需改动。
 	// 新波次僵尸生成时(CreateZombie)对其 body/头盔/护盾血量整体乘此系数；读档(CreateZombieWithID)不乘，避免二次叠加。
-	float GetZombieHpMultiplier() const {
-		float multiplier = 1.0f;
+	double GetZombieHpMultiplier() const {
+		double multiplier = 1.0;
 		if (mIsSurvival)
-			multiplier *= (1.0f + SURVIVAL_HP_GROWTH * static_cast<float>(mSurvivalRound - 1));
+			multiplier *= (1.0 + SURVIVAL_HP_GROWTH * static_cast<double>(mSurvivalRound - 1));
 		return multiplier;
 	}
